Preparing as secondary
The main reason sailfish doesn't boot as a secondary is the lack of finding it's original /system of cm12.1. Sailfish wasn't meant for a secondary ROM. So what it does on boot is always mounting mmcblk0p23 via system.mount. That block however, always contains /system of the primary ROM. For the mismatch of expected and found /system, HAL doesn't get initialized in the way it should, and that's why we got no change from the splash screen when we boot exactly the same sailfish as secondary. That's why we're tweaking the way Sailfish mounts /system here. MultiROM also, is made for an universal /system configuration. Not the chroot type config Sailfish uses. So it always mounts the secondary ROMs /system to the root /system. We're  binding those two and deleting */system.mount's for the sake of booting as a secondary ROM.